Adding the other test cases

+++ const.h
@@ -23,7 +23,7 @@ using yakl::memset;
   #define NGLL 3
 #endif
 
-typedef float real;
+typedef double real;
 
 YAKL_INLINE real constexpr operator"" _fp( long double x ) {
   return static_cast<real>(x);
